# use Math.imul