--- a/.gitignore
+++ b/.gitignore
@@ -16,3 +16,4 @@ node_modules
 npm-debug.log
 coverage/
 lib/binding
+.idea