--- a/lib/murmurhash.js
+++ b/lib/murmurhash.js
@@ -14,6 +14,7 @@
 
 var murmur = require('./murmur');
 
+
 /**
  * Murmur hash v2
  *