Math.imul is new class method of es6.

```
murmurhash should be 335538535: 335538535
murmurhash utf8 should be 3300337389: 3300337389
murmurhash with custom key should be 4040304760: 4040304760
murmurhash('hello 中国') should be 1248731102: 1248731102

  node version: v4.2.1, date: Sat Nov 14 2015 21:59:18 GMT+0800 (CST)
use Math.imul
  Starting...
  2 tests completed.

  murmurhash(new Buffer(ascii)) x 20,148,408 ops/sec ±0.74% (97 runs sampled)
  murmurhash(new Buffer(utf8))  x 12,077,352 ops/sec ±0.86% (100 runs sampled)
```

almost 2.5x faster.