@alsotang 2015-11-14
for speed consideration, dont abstract imul to a function. just use two separate files.