@alsotang 2015-11-14
哈哈，我还用 `buf.writeUInt8(key[i + 2], 0), buf.readInt32BE(0)` 来代替我们那个 

```js
// @zhangzifa: c语言中，大于127的char, 在转成 int 时, 表示成0xffffffxx的形式
var NEGTIVE_MAP = [0x0,  0x1,  0x2,  0x3,  0x4,  0x5,  0x6,  0x7,  0x8,  0x9,  0xa,  0xb,...]
```

结果性能下降了。。