@alsotang 2015-11-14
用 Buffer 的思路不是很对。

```js
case 3: k1 ^= (key.charCodeAt(i + 2) & 0xff) << 16;
		case 2: k1 ^= (key.charCodeAt(i + 1) & 0xff) << 8;
		case 1: k1 ^= (key.charCodeAt(i) & 0xff);
```

看到一句这样的，其实也蛮好。不一定比我们现在的查表慢。我试试。