@alsotang 2015-11-14
用上 `& 0xff` 快了，已经在 pr 中实现。