@alsotang 2015-11-14
before Math.imul
```js
alsotang ~/Dropbox/code/node-murmurhash (imul *) $ node test/benchmark.js

murmurhash should be 335538535: 335538535
murmurhash utf8 should be 3300337389: 9505949
murmurhash with custom key should be 4040304760: 2934831275
murmurhash('hello 中国') should be 1248731102: 1248731102

  node version: v4.2.1, date: Sun Nov 15 2015 01:02:17 GMT+0800 (CST)
  Starting...
  2 tests completed.

  murmurhash(new Buffer(ascii)) x 8,867,064 ops/sec ±0.99% (98 runs sampled)
  murmurhash(new Buffer(utf8))  x 5,027,139 ops/sec ±0.87% (97 runs sampled)
```


after Math.imul
```js
alsotang ~/Dropbox/code/node-murmurhash (imul) $ node test/benchmark.js

murmurhash should be 335538535: 335538535
murmurhash utf8 should be 3300337389: 9505949
murmurhash with custom key should be 4040304760: 2934831275
murmurhash('hello 中国') should be 1248731102: 1248731102

  node version: v4.2.1, date: Sun Nov 15 2015 01:01:49 GMT+0800 (CST)
  Starting...
  2 tests completed.

  murmurhash(new Buffer(ascii)) x 21,164,676 ops/sec ±0.85% (98 runs sampled)
  murmurhash(new Buffer(utf8))  x 13,926,544 ops/sec ±0.59% (99 runs sampled)
```