@alsotang 2015-11-14
```js
    // k = (key[i] |
    //   (key[i + 1] << 8) |
    //   (key[i + 2] << 16) |
    //   (key[i + 3] << 24));
    k = key.readUInt32LE(i)
```
这个地方，使用 Buffer 的方法反而慢了