@alsotang 2015-11-14
v8 在函数热缓存这块比我当时测的时候要好，现在合并成一个文件了。旧版本的性能没有下降。